% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_database.R
\name{add_file_info}
\alias{add_file_info}
\title{Add File Information From File Database}
\usage{
add_file_info(data)
}
\arguments{
\item{data}{data frame with column \code{file_id} containing file identifiers
and with an attribute \code{file_db} containing a "file database" as
created by \code{to_file_database}}
}
\value{
data frame \code{data} with additional columns \code{folder_path}
  and \code{file_name}
}
\description{
Add File Information From File Database
}
\examples{
# Define some paths
paths <- c(
  "/very/long/path/very_long_file_name_1",
  "/very/long/path/very_long_file_name_2",
  "/very/long/path/very_long_file_name_3"
)

# Create a "file database" from the paths
file_db <- kwb.file::to_file_database(paths, remove_common_base = FALSE)

# Create a data frame that relates some information to the files.
# Use the file identifier instead of the full name to keep the data clean
(df <- kwb.utils::noFactorDataFrame(
  file_id = file_db$files$file_id, 
  value = seq_along(paths)
))

# Store the file database in the attribute "file_db"
df <- structure(df, file_db = file_db)

# Restore the full file paths
add_file_info(df)

}
