% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{copy_files_to_target_dir}
\alias{copy_files_to_target_dir}
\title{Copy Files to Flat Structure}
\usage{
copy_files_to_target_dir(from_paths, target_dir, target_files)
}
\arguments{
\item{from_paths}{paths to the files to be copied}

\item{target_dir}{path to the target directory}

\item{target_files}{relative paths to the target files, relative to
\code{target_dir}}
}
\description{
Calls \code{file.copy} under the hood but gives a message about the indices
and paths of the files that could not be copied.
}
\examples{
root <- system.file(package = "kwb.file")

relative_paths <- dir(root, recursive = TRUE)

# The original files are in root or in different subfolders
relative_paths

# Create a temporary target folder
target_dir <- kwb.utils::createDirectory(file.path(tempdir(), "target"))

# Copy all files into one target folder without subfolders
from_paths <- file.path(root, relative_paths)
to_paths <- basename(from_paths)

# Make sure that the target file names contain no duplicates, otherwise
# an error is raised
to_paths <- kwb.utils::makeUnique(to_paths, warn = FALSE)

# Copy the files
copy_files_to_target_dir(from_paths, target_dir, to_paths)

# Look at the result
dir(target_dir, recursive = TRUE)

}
