% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{dir_full}
\alias{dir_full}
\title{Helper function to return full paths}
\usage{
dir_full(...)
}
\arguments{
\item{\dots}{arguments passed to \code{\link{dir}}}
}
\description{
This function provides a shortcut to \code{dir(..., full.names = TRUE)}
}
\examples{
dir_full(system.file(package = "kwb.file"))

}
