% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{dir_full_recursive_xml}
\alias{dir_full_recursive_xml}
\title{Get Full Paths to all XML files Below a Root Folder}
\usage{
dir_full_recursive_xml(root)
}
\arguments{
\item{root}{path to root folder}
}
\value{
vector of character
}
\description{
Get Full Paths to all XML files Below a Root Folder
}
