% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{get_download_dir}
\alias{get_download_dir}
\title{Get Default Download Directory}
\usage{
get_download_dir()
}
\value{
assumed default download directory on the user's computer (vector of
  character of length one)
}
\description{
Get Default Download Directory
}
\examples{
dir_full(get_download_dir())

}
