% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_metadata.R
\name{read_file_metadata}
\alias{read_file_metadata}
\title{Read File Metadata from YAML-File}
\usage{
read_file_metadata(yaml_file, file_encoding = "UTF-8",
  out_class = c("data.frame", "list")[1])
}
\arguments{
\item{yaml_file}{path to YAML-File containing file metadata (as saved with
\code{kwb.file:::write_file_info_to_yaml_file})}

\item{file_encoding}{passed to argument \code{fileEncoding} of
\code{\link[yaml]{read_yaml}}}

\item{out_class}{one of "data.frame", "list"}
}
\value{
depending on \code{out_class}, either a data frame with the following
  columns or a list with the following elements is returned: 
  \describe{
    \item{file_id}{clean file name given to the original file for simpler
      access}
    \item{original_name}{original file name given by data provider}
    \item{original_folder}{original path to folder in which file was
      provided.}
  }
}
\description{
Read File Metadata from YAML-File
}
