% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_common_root.R
\name{remove_common_root}
\alias{remove_common_root}
\title{Remove the Common Root Parts}
\usage{
remove_common_root(x, n_keep = 1L, dbg = TRUE)
}
\arguments{
\item{x}{list of vectors of character as returned by
\code{\link[base]{strsplit}} or a vector of character.}

\item{n_keep}{minimum number of segments to be kept in any case in the 
returned relative paths. For example, two paths "a" and "a/b" have the 
common root "a". Removing this root would result in relative paths 
"" and "b". As this is not useful, \code{n_keep} is \code{1} by default,
making sure that all paths keep at least one segment (segment "a") in the
example.}

\item{dbg}{if \code{TRUE} debug messages are shown}
}
\description{
Remove the Common Root Parts
}
\examples{
# Split paths at the slashes
absparts <- strsplit(c("a/b/c", "a/b/d", "a/b/e/f/g", "a/b/hi"), "/")

# Remove the common parts of the paths
relparts <- remove_common_root(absparts)
relparts

# The extracted root is returned in attribute "root"
attr(relparts, "root")

}
