% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{split_into_dir_and_file}
\alias{split_into_dir_and_file}
\title{"Split Full Paths into Directory Path and Filename"}
\usage{
split_into_dir_and_file(paths)
}
\arguments{
\item{paths}{vector of character representing full file paths}
}
\value{
data frame with columns \code{directory} and \code{file}
}
\description{
"Split Full Paths into Directory Path and Filename"
}
\examples{
split_into_dir_and_file(c("path/to/file-1", "path/to/file-2"))
}
