% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{split_paths}
\alias{split_paths}
\title{Split Full Paths at Slashes into Parts}
\usage{
split_paths(paths, dbg = TRUE)
}
\arguments{
\item{paths}{vector of character representing full file paths}

\item{dbg}{if \code{TRUE} (default), a debug message is shown}
}
\description{
Split Full Paths at Slashes into Parts
}
\examples{
segments <- split_paths(c("path/to/file-1", "path/to/file-2"))
segments
}
