% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_database.R
\name{to_file_database}
\alias{to_file_database}
\title{Create Two Table Relational Database From Paths}
\usage{
to_file_database(files, remove_common_base = TRUE)
}
\arguments{
\item{files}{vector of file paths}

\item{remove_common_base}{if \code{TRUE} (default) the common root of all
\code{files} is removed before creating the database}
}
\value{
list of two data frames, \code{files} and \code{folders}
}
\description{
From a vector of given file paths, this function generates short and unique 
identifiers for files and folders. The assignements between identifiers and 
original paths are stored in two data frames, \code{files} and \code{folders}
that are returned.
}
\examples{
paths <- c(
  "very_long/very_ugly_path/even with spaces.doc",
  "very_long/very_ugly_path/even with spaces.docx"
)

to_file_database(paths)
to_file_database(paths, remove_common_base = FALSE)

}
